fix the second.py second input by using float operator

+++ second.py
@@ -1,6 +1,6 @@
 num1 = float(input("Enter the first number : "))
 operator = str(input("Select any operation, =,-,*,/ : "))
-num2 = int(input("Enter the second number : "))
+num2 = float(input("Enter the second number : "))
 
 if operator == "+" : print("Result :",num1+num2)
 elif operator == "-" : print("Result :",num1-num2)
